+++ 2021/day09/lib.h
@@ -6,8 +6,8 @@
 
 namespace day09
 {
-    uint32_t run_part1(const std::vector<std::string> &input);
-    uint32_t run_part2(const std::vector<std::string> &input);
+    int64_t run_part1(const std::vector<std::string> &input);
+    int64_t run_part2(const std::vector<std::string> &input);
 }
 
 #endif
